Extracting time taken from wrapper
