+++ src/utils/testconfig.ini
@@ -18,4 +18,4 @@ formatter=simpleFormatter
 args=(__import__("datetime").datetime.now().strftime(f"src/utils/{__name__}_%%Y-%%m-%%d_%%H-%%M-%%S.log"),)
 
 [formatter_simpleFormatter]
-format=%(asctime)s %(name)s - %(levelname)s:%(message)s
+format=%(asctime)s %(name)s - %(funcName)s - %(levelname)s:%(message)s
